+++ backend/canisters/group/c2c_client/src/lib.rs
@@ -2,7 +2,10 @@ use canister_client::{generate_c2c_call, generate_candid_c2c_call};
 use group_canister::*;
 
 // Queries
+generate_c2c_call!(c2c_events);
+generate_c2c_call!(c2c_events_by_index);
 generate_c2c_call!(c2c_events_internal);
+generate_c2c_call!(c2c_events_window);
 generate_c2c_call!(c2c_name_and_members);
 generate_c2c_call!(c2c_summary);
 generate_c2c_call!(c2c_summary_updates);