+++ backend/libraries/types/src/chat_summary.rs
@@ -1,7 +1,7 @@
 use crate::{
-    AccessGate, BuildVersion, CanisterId, ChatId, EventIndex, EventWrapper, FrozenGroupInfo, GroupMember, GroupPermissions,
-    GroupRole, HydratedMention, Message, MessageIndex, Milliseconds, OptionUpdate, TimestampMillis, UserId, Version,
-    MAX_RETURNED_MENTIONS,
+    local_user_index_canister_id, AccessGate, BuildVersion, CanisterId, ChatId, EventIndex, EventWrapper, FrozenGroupInfo,
+    GroupMember, GroupPermissions, GroupRole, HydratedMention, Message, MessageIndex, Milliseconds, OptionUpdate,
+    TimestampMillis, UserId, Version, MAX_RETURNED_MENTIONS,
 };
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
@@ -34,8 +34,10 @@ impl DirectChatSummary {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "GroupChatSummaryPrevious")]
 pub struct GroupChatSummary {
     pub chat_id: ChatId,
+    pub local_user_index_canister_id: CanisterId,
     pub last_updated: TimestampMillis,
     pub name: String,
     pub description: String,
@@ -70,6 +72,83 @@ pub struct GroupChatSummary {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct GroupChatSummaryPrevious {
+    pub chat_id: ChatId,
+    pub local_user_index_canister_id: CanisterId,
+    pub last_updated: TimestampMillis,
+    pub name: String,
+    pub description: String,
+    pub subtype: Option<GroupSubtype>,
+    pub avatar_id: Option<u128>,
+    pub is_public: bool,
+    pub history_visible_to_new_joiners: bool,
+    pub min_visible_event_index: EventIndex,
+    pub min_visible_message_index: MessageIndex,
+    pub latest_message: Option<EventWrapper<Message>>,
+    pub latest_event_index: EventIndex,
+    pub latest_message_index: Option<MessageIndex>,
+    pub joined: TimestampMillis,
+    pub read_by_me_up_to: Option<MessageIndex>,
+    pub notifications_muted: bool,
+    pub participant_count: u32,
+    pub role: GroupRole,
+    pub mentions: Vec<HydratedMention>,
+    pub wasm_version: BuildVersion,
+    pub permissions_v2: GroupPermissions,
+    pub metrics: ChatMetrics,
+    pub my_metrics: ChatMetrics,
+    pub latest_threads: Vec<ThreadSyncDetails>,
+    pub archived: bool,
+    pub frozen: Option<FrozenGroupInfo>,
+    pub date_last_pinned: Option<TimestampMillis>,
+    pub date_read_pinned: Option<TimestampMillis>,
+    pub events_ttl: Option<Milliseconds>,
+    pub events_ttl_last_updated: TimestampMillis,
+    pub gate: Option<AccessGate>,
+    pub rules_accepted: bool,
+}
+
+impl From<GroupChatSummaryPrevious> for GroupChatSummary {
+    fn from(value: GroupChatSummaryPrevious) -> Self {
+        GroupChatSummary {
+            chat_id: value.chat_id,
+            local_user_index_canister_id: local_user_index_canister_id(value.chat_id.into()),
+            last_updated: value.last_updated,
+            name: value.name,
+            description: value.description,
+            subtype: value.subtype,
+            avatar_id: value.avatar_id,
+            is_public: value.is_public,
+            history_visible_to_new_joiners: value.history_visible_to_new_joiners,
+            min_visible_event_index: value.min_visible_event_index,
+            min_visible_message_index: value.min_visible_message_index,
+            latest_message: value.latest_message,
+            latest_event_index: value.latest_event_index,
+            latest_message_index: value.latest_message_index,
+            joined: value.joined,
+            read_by_me_up_to: value.read_by_me_up_to,
+            notifications_muted: value.notifications_muted,
+            participant_count: value.participant_count,
+            role: value.role,
+            mentions: value.mentions,
+            wasm_version: value.wasm_version,
+            permissions_v2: value.permissions_v2,
+            metrics: value.metrics,
+            my_metrics: value.my_metrics,
+            latest_threads: value.latest_threads,
+            archived: value.archived,
+            frozen: value.frozen,
+            date_last_pinned: value.date_last_pinned,
+            date_read_pinned: value.date_read_pinned,
+            events_ttl: value.events_ttl,
+            events_ttl_last_updated: value.events_ttl_last_updated,
+            gate: value.gate,
+            rules_accepted: value.rules_accepted,
+        }
+    }
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct DirectChatSummaryUpdates {
     pub chat_id: ChatId,
     pub last_updated: TimestampMillis,
@@ -91,8 +170,10 @@ pub struct DirectChatSummaryUpdates {
 // which is causing unnecessarily coupling. We should use separate types for these use cases.
 // For instance we only need to return history_visible_to_new_joiners and is_public from group::public_summary
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "PublicGroupSummaryPrevious")]
 pub struct PublicGroupSummary {
     pub chat_id: ChatId,
+    pub local_user_index_canister_id: CanisterId,
     pub last_updated: TimestampMillis,
     pub name: String,
     pub description: String,
@@ -112,7 +193,89 @@ pub struct PublicGroupSummary {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct PublicGroupSummaryPrevious {
+    pub chat_id: ChatId,
+    pub last_updated: TimestampMillis,
+    pub name: String,
+    pub description: String,
+    pub subtype: Option<GroupSubtype>,
+    pub history_visible_to_new_joiners: bool,
+    pub avatar_id: Option<u128>,
+    pub latest_message: Option<EventWrapper<Message>>,
+    pub latest_event_index: EventIndex,
+    pub latest_message_index: Option<MessageIndex>,
+    pub participant_count: u32,
+    pub wasm_version: BuildVersion,
+    pub is_public: bool,
+    pub frozen: Option<FrozenGroupInfo>,
+    pub events_ttl: Option<Milliseconds>,
+    pub events_ttl_last_updated: TimestampMillis,
+    pub gate: Option<AccessGate>,
+}
+
+impl From<PublicGroupSummaryPrevious> for PublicGroupSummary {
+    fn from(value: PublicGroupSummaryPrevious) -> Self {
+        PublicGroupSummary {
+            chat_id: value.chat_id,
+            local_user_index_canister_id: local_user_index_canister_id(value.chat_id.into()),
+            last_updated: value.last_updated,
+            name: value.name,
+            description: value.description,
+            subtype: value.subtype,
+            history_visible_to_new_joiners: value.history_visible_to_new_joiners,
+            avatar_id: value.avatar_id,
+            latest_message: value.latest_message,
+            latest_event_index: value.latest_event_index,
+            latest_message_index: value.latest_message_index,
+            participant_count: value.participant_count,
+            wasm_version: value.wasm_version,
+            is_public: value.is_public,
+            frozen: value.frozen,
+            events_ttl: value.events_ttl,
+            events_ttl_last_updated: value.events_ttl_last_updated,
+            gate: value.gate,
+        }
+    }
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "GroupCanisterGroupChatSummaryPrevious")]
 pub struct GroupCanisterGroupChatSummary {
+    pub chat_id: ChatId,
+    pub local_user_index_canister_id: CanisterId,
+    pub last_updated: TimestampMillis,
+    pub name: String,
+    pub description: String,
+    pub subtype: Option<GroupSubtype>,
+    pub avatar_id: Option<u128>,
+    pub is_public: bool,
+    pub history_visible_to_new_joiners: bool,
+    pub min_visible_event_index: EventIndex,
+    pub min_visible_message_index: MessageIndex,
+    pub latest_message: Option<EventWrapper<Message>>,
+    pub latest_event_index: EventIndex,
+    pub latest_message_index: Option<MessageIndex>,
+    pub joined: TimestampMillis,
+    pub participant_count: u32,
+    pub role: GroupRole,
+    pub mentions: Vec<HydratedMention>,
+    pub wasm_version: BuildVersion,
+    pub permissions_v2: GroupPermissions,
+    pub notifications_muted: bool,
+    pub metrics: ChatMetrics,
+    pub my_metrics: ChatMetrics,
+    pub latest_threads: Vec<GroupCanisterThreadDetails>,
+    pub frozen: Option<FrozenGroupInfo>,
+    pub date_last_pinned: Option<TimestampMillis>,
+    pub events_ttl: Option<Milliseconds>,
+    pub events_ttl_last_updated: TimestampMillis,
+    pub gate: Option<AccessGate>,
+    pub rules_accepted: bool,
+    pub membership: Option<GroupMembership>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct GroupCanisterGroupChatSummaryPrevious {
     pub chat_id: ChatId,
     pub last_updated: TimestampMillis,
     pub name: String,
@@ -143,6 +306,44 @@ pub struct GroupCanisterGroupChatSummary {
     pub gate: Option<AccessGate>,
     pub rules_accepted: bool,
     pub membership: Option<GroupMembership>,
+}
+
+impl From<GroupCanisterGroupChatSummaryPrevious> for GroupCanisterGroupChatSummary {
+    fn from(value: GroupCanisterGroupChatSummaryPrevious) -> Self {
+        GroupCanisterGroupChatSummary {
+            chat_id: value.chat_id,
+            local_user_index_canister_id: local_user_index_canister_id(value.chat_id.into()),
+            last_updated: value.last_updated,
+            name: value.name,
+            description: value.description,
+            subtype: value.subtype,
+            avatar_id: value.avatar_id,
+            is_public: value.is_public,
+            history_visible_to_new_joiners: value.history_visible_to_new_joiners,
+            min_visible_event_index: value.min_visible_event_index,
+            min_visible_message_index: value.min_visible_message_index,
+            latest_message: value.latest_message,
+            latest_event_index: value.latest_event_index,
+            latest_message_index: value.latest_message_index,
+            joined: value.joined,
+            participant_count: value.participant_count,
+            role: value.role,
+            mentions: value.mentions,
+            wasm_version: value.wasm_version,
+            permissions_v2: value.permissions_v2,
+            notifications_muted: value.notifications_muted,
+            metrics: value.metrics,
+            my_metrics: value.my_metrics,
+            latest_threads: value.latest_threads,
+            frozen: value.frozen,
+            date_last_pinned: value.date_last_pinned,
+            events_ttl: value.events_ttl,
+            events_ttl_last_updated: value.events_ttl_last_updated,
+            gate: value.gate,
+            rules_accepted: value.rules_accepted,
+            membership: value.membership,
+        }
+    }
 }
 
 impl GroupCanisterGroupChatSummary {
@@ -186,6 +387,7 @@ impl GroupCanisterGroupChatSummary {
 
         GroupCanisterGroupChatSummary {
             chat_id: self.chat_id,
+            local_user_index_canister_id: self.local_user_index_canister_id,
             last_updated: updates.last_updated,
             name: updates.name.unwrap_or(self.name),
             description: updates.description.unwrap_or(self.description),