+++ frontend/app/src/components/home/profile/Accounts.svelte
@@ -35,6 +35,8 @@
     $: cryptoLookup = client.cryptoLookup;
     $: cryptoBalance = client.cryptoBalance;
     $: accounts = buildAccountsList($cryptoLookup, $cryptoBalance);
+    $: nervousSystemLookup = client.nervousSystemLookup;
+    $: snsLedgers = new Set<string>(Object.values($nervousSystemLookup).map((ns) => ns.ledgerCanisterId));
 
     $: {
         zeroCount = accounts.filter((a) => a.zero).length;
@@ -167,7 +169,7 @@
                                         slot="icon" />
                                     <div slot="text">{$_("cryptoAccount.receive")}</div>
                                 </MenuItem>
-                                {#if !["ckbtc", "icp"].includes(token.symbol.toLowerCase())}
+                                {#if snsLedgers.has(token.ledger)}
                                     <MenuItem on:click={() => (transactionsFor = token)}>
                                         <ViewList
                                             size={$iconSize}